Updated training wheel in both car.inos. Sklearn/Autopilot stuff in Pycharm Files

+++ app/DataController.java
@@ -3,6 +3,8 @@ package app;
 import com.opencsv.CSVReader;
 import com.opencsv.CSVWriter;
 
+import app.DataPacket;
+
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
@@ -42,23 +44,23 @@ public class DataController
 
 			while ((element = reader.readNext()) != null)
 			{
-				double leftUltrasonic = Double.parseDouble(element[0]);
-				double upperLeftUltrasonic = Double.parseDouble(element[1]);
-				double middleUltrasonic = Double.parseDouble(element[2]);
-				double upperRightUltrasonic = Double.parseDouble(element[3]);
-				double rightUltrasonic = Double.parseDouble(element[4]);
+				int leftUltrasonic = Integer.parseInt(element[0]);
+				int upperLeftUltrasonic = Integer.parseInt(element[1]);
+				int middleUltrasonic = Integer.parseInt(element[2]);
+				int upperRightUltrasonic = Integer.parseInt(element[3]);
+				int rightUltrasonic = Integer.parseInt(element[4]);
 				int db1 = Integer.parseInt(element[5]);
 				int db2 = Integer.parseInt(element[6]);
 				int db3 = Integer.parseInt(element[7]);
-				double xAccel = Double.parseDouble(element[8]);
-				double yAccel = Double.parseDouble(element[9]);
-				double zAccel = Double.parseDouble(element[10]);
-				double xGyro = Double.parseDouble(element[11]);
-				double yGyro = Double.parseDouble(element[12]);
-				double zGyro = Double.parseDouble(element[13]);
-				double xMag = Double.parseDouble(element[14]);
-				double yMag = Double.parseDouble(element[15]);
-				double zMag = Double.parseDouble(element[16]);
+				int xAccel = Integer.parseInt(element[8]);
+				int yAccel = Integer.parseInt(element[9]);
+				int zAccel = Integer.parseInt(element[10]);
+				int xGyro = Integer.parseInt(element[11]);
+				int yGyro = Integer.parseInt(element[12]);
+				int zGyro = Integer.parseInt(element[13]);
+				int xMag = Integer.parseInt(element[14]);
+				int yMag = Integer.parseInt(element[15]);
+				int zMag = Integer.parseInt(element[16]);
 				int servoAngle = Integer.parseInt(element[17]);
 				int state = Integer.parseInt(element[18]);
 
@@ -163,23 +165,23 @@ public class DataController
 
 			PreparedStatement pstmt = c.prepareStatement(sql);
 
-			pstmt.setDouble(1, packet.getLeftUltrasonic());
-			pstmt.setDouble(2, packet.getUpperLeftUltrasonic());
-			pstmt.setDouble(3, packet.getMiddleUltrasonic());
-			pstmt.setDouble(4, packet.getUpperRightUltrasonic());
-			pstmt.setDouble(5, packet.getRightUltrasonic());
+			pstmt.setInt(1, packet.getLeftUltrasonic());
+			pstmt.setInt(2, packet.getUpperLeftUltrasonic());
+			pstmt.setInt(3, packet.getMiddleUltrasonic());
+			pstmt.setInt(4, packet.getUpperRightUltrasonic());
+			pstmt.setInt(5, packet.getRightUltrasonic());
 			pstmt.setInt(6, packet.getDb1());
 			pstmt.setInt(7, packet.getDb2());
 			pstmt.setInt(8, packet.getDb3());
-			pstmt.setDouble(9, packet.getxAccel());
-			pstmt.setDouble(10, packet.getyAccel());
-			pstmt.setDouble(11, packet.getzAccel());
-			pstmt.setDouble(12, packet.getxGyro());
-			pstmt.setDouble(13, packet.getyGyro());
-			pstmt.setDouble(14, packet.getzGyro());
-			pstmt.setDouble(15, packet.getxMag());
-			pstmt.setDouble(16, packet.getyMag());
-			pstmt.setDouble(17, packet.getzMag());
+			pstmt.setInt(9, packet.getxAccel());
+			pstmt.setInt(10, packet.getyAccel());
+			pstmt.setInt(11, packet.getzAccel());
+			pstmt.setInt(12, packet.getxGyro());
+			pstmt.setInt(13, packet.getyGyro());
+			pstmt.setInt(14, packet.getzGyro());
+			pstmt.setInt(15, packet.getxMag());
+			pstmt.setInt(16, packet.getyMag());
+			pstmt.setInt(17, packet.getzMag());
 			pstmt.setInt(18,  packet.getServoAngle());
 			pstmt.setInt(19, packet.getState());
 			pstmt.executeUpdate();
@@ -226,9 +228,9 @@ public class DataController
 
 			while (rs.next())
 			{
-				DataPacket packet = new DataPacket(rs.getDouble("leftUltrasonic"), rs.getDouble("upperLeftUltrasonic"), rs.getDouble("middleUltrasonic"), rs.getDouble("upperRightUltrasonic"), rs.getDouble("rightUltrasonic"), rs.getInt("db1"), rs.getInt("db2"), rs.getInt("db3"),
-						rs.getDouble("xAccel"), rs.getDouble("yAccel"), rs.getDouble("zAccel"), rs.getDouble("xGyro"), rs.getDouble("yGyro"), rs.getDouble("zGyro"),
-						rs.getDouble("xMag"), rs.getDouble("yMag"), rs.getDouble("zMag"), rs.getInt("servoAngle"), rs.getInt("state"));
+				DataPacket packet = new DataPacket(rs.getInt("leftUltrasonic"), rs.getInt("upperLeftUltrasonic"), rs.getInt("middleUltrasonic"), rs.getInt("upperRightUltrasonic"), rs.getInt("rightUltrasonic"), rs.getInt("db1"), rs.getInt("db2"), rs.getInt("db3"),
+						rs.getInt("xAccel"), rs.getInt("yAccel"), rs.getInt("zAccel"), rs.getInt("xGyro"), rs.getInt("yGyro"), rs.getInt("zGyro"),
+						rs.getInt("xMag"), rs.getInt("yMag"), rs.getInt("zMag"), rs.getInt("servoAngle"), rs.getInt("state"));
 
 				data.add(packet);
 			}
