+++ app/DataPacket.java
@@ -5,23 +5,23 @@ import java.text.DecimalFormat;
 import java.text.NumberFormat;
 
 public class DataPacket {
-	public double leftUltrasonic;
-	public double upperLeftUltrasonic;
-	public double middleUltrasonic;
-	public double upperRightUltrasonic;
-	public double rightUltrasonic;
+	public int leftUltrasonic;
+	public int upperLeftUltrasonic;
+	public int middleUltrasonic;
+	public int upperRightUltrasonic;
+	public int rightUltrasonic;
 	public int db1;
 	public int db2;
 	public int db3;
-	public double xAccel;
-	public double yAccel;
-	public double zAccel;
-	public double xGyro;
-	public double yGyro;
-	public double zGyro;
-	public double xMag;
-	public double yMag;
-	public double zMag;
+	public int xAccel;
+	public int yAccel;
+	public int zAccel;
+	public int xGyro;
+	public int yGyro;
+	public int zGyro;
+	public int xMag;
+	public int yMag;
+	public int zMag;
 	public int servoAngle;
 	public int state;
 
@@ -45,9 +45,9 @@ public class DataPacket {
 
 	}
 
-	public DataPacket(double leftUltrasonic, double upperLeftUltrasonic, double middleUltrasonic,
-			double upperRightUltrasonic, double rightUltrasonic, int db1, int db2, int db3, double xAccel, double yAccel,
-			double zAccel, double xGyro, double yGyro, double zGyro, double xMag, double yMag, double zMag,
+	public DataPacket(int leftUltrasonic, int upperLeftUltrasonic, int middleUltrasonic,
+			int upperRightUltrasonic, int rightUltrasonic, int db1, int db2, int db3, int xAccel, int yAccel,
+			int zAccel, int xGyro, int yGyro, int zGyro, int xMag, int yMag, int zMag,
 			int servoAngle, int state) {
 		this.leftUltrasonic = leftUltrasonic;
 		this.upperLeftUltrasonic = upperLeftUltrasonic;
@@ -70,43 +70,43 @@ public class DataPacket {
 		this.state = state;
 	}
 
-	public double getLeftUltrasonic() {
+	public int getLeftUltrasonic() {
 		return leftUltrasonic;
 	}
 
-	public void setLeftUltrasonic(double leftUltrasonic) {
+	public void setLeftUltrasonic(int leftUltrasonic) {
 		this.leftUltrasonic = leftUltrasonic;
 	}
 
-	public double getUpperLeftUltrasonic() {
+	public int getUpperLeftUltrasonic() {
 		return upperLeftUltrasonic;
 	}
 
-	public void setUpperLeftUltrasonic(double upperLeftUltrasonic) {
+	public void setUpperLeftUltrasonic(int upperLeftUltrasonic) {
 		this.upperLeftUltrasonic = upperLeftUltrasonic;
 	}
 
-	public double getMiddleUltrasonic() {
+	public int getMiddleUltrasonic() {
 		return middleUltrasonic;
 	}
 
-	public void setMiddleUltrasonic(double middleUltrasonic) {
+	public void setMiddleUltrasonic(int middleUltrasonic) {
 		this.middleUltrasonic = middleUltrasonic;
 	}
 
-	public double getUpperRightUltrasonic() {
+	public int getUpperRightUltrasonic() {
 		return upperRightUltrasonic;
 	}
 
-	public void setUpperRightUltrasonic(double upperRightUltrasonic) {
+	public void setUpperRightUltrasonic(int upperRightUltrasonic) {
 		this.upperRightUltrasonic = upperRightUltrasonic;
 	}
 
-	public double getRightUltrasonic() {
+	public int getRightUltrasonic() {
 		return rightUltrasonic;
 	}
 
-	public void setRightUltrasonic(double rightUltrasonic) {
+	public void setRightUltrasonic(int rightUltrasonic) {
 		this.rightUltrasonic = rightUltrasonic;
 	}
 
@@ -134,75 +134,75 @@ public class DataPacket {
 		this.db3 = db3;
 	}
 
-	public double getxAccel() {
+	public int getxAccel() {
 		return xAccel;
 	}
 
-	public void setxAccel(double xAccel) {
+	public void setxAccel(int xAccel) {
 		this.xAccel = xAccel;
 	}
 
-	public double getyAccel() {
+	public int getyAccel() {
 		return yAccel;
 	}
 
-	public void setyAccel(double yAccel) {
+	public void setyAccel(int yAccel) {
 		this.yAccel = yAccel;
 	}
 
-	public double getzAccel() {
+	public int getzAccel() {
 		return zAccel;
 	}
 
-	public void setzAccel(double zAccel) {
+	public void setzAccel(int zAccel) {
 		this.zAccel = zAccel;
 	}
 
-	public double getxGyro() {
+	public int getxGyro() {
 		return xGyro;
 	}
 
-	public void setxGyro(double xGyro) {
+	public void setxGyro(int xGyro) {
 		this.xGyro = xGyro;
 	}
 
-	public double getyGyro() {
+	public int getyGyro() {
 		return yGyro;
 	}
 
-	public void setyGyro(double yGyro) {
+	public void setyGyro(int yGyro) {
 		this.yGyro = yGyro;
 	}
 
-	public double getzGyro() {
+	public int getzGyro() {
 		return zGyro;
 	}
 
-	public void setzGyro(double zGyro) {
+	public void setzGyro(int zGyro) {
 		this.zGyro = zGyro;
 	}
 
-	public double getxMag() {
+	public int getxMag() {
 		return xMag;
 	}
 
-	public void setxMag(double xMag) {
+	public void setxMag(int xMag) {
 		this.xMag = xMag;
 	}
 
-	public double getyMag() {
+	public int getyMag() {
 		return yMag;
 	}
 
-	public void setyMag(double yMag) {
+	public void setyMag(int yMag) {
 		this.yMag = yMag;
 	}
 
-	public double getzMag() {
+	public int getzMag() {
 		return zMag;
 	}
 
-	public void setzMag(double zMag) {
+	public void setzMag(int zMag) {
 		this.zMag = zMag;
 	}
